+++ assembly-syntax/src/ast/attribute/meta/kv.rs
@@ -1,14 +1,21 @@
 use alloc::collections::BTreeMap;
 use core::borrow::Borrow;
 
+use miden_core::utils::{
+    ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable,
+};
 use miden_debug_types::{SourceSpan, Spanned};
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 
 use super::MetaExpr;
 use crate::ast::Ident;
 
 /// Represents the metadata of a key-value [crate::ast::Attribute], i.e. `@props(key = value)`
-#[derive(Clone)]
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct MetaKeyValue {
+    #[cfg_attr(feature = "serde", serde(skip, default))]
     pub span: SourceSpan,
     /// The name of the key-value dictionary
     pub name: Ident,
@@ -133,5 +140,38 @@ impl core::hash::Hash for MetaKeyValue {
     fn hash<H: core::hash::Hasher>(&self, state: &mut H) {
         self.name.hash(state);
         self.items.hash(state);
+    }
+}
+
+#[cfg(feature = "arbitrary")]
+impl proptest::arbitrary::Arbitrary for MetaKeyValue {
+    type Parameters = ();
+
+    fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
+        use proptest::{arbitrary::any, strategy::Strategy};
+
+        let name = any::<Ident>();
+        let items = proptest::collection::btree_map(any::<Ident>(), any::<MetaExpr>(), 1..3);
+        (name, items)
+            .prop_map(|(name, items)| Self { span: SourceSpan::UNKNOWN, name, items })
+            .boxed()
+    }
+
+    type Strategy = proptest::prelude::BoxedStrategy<Self>;
+}
+
+impl Serializable for MetaKeyValue {
+    fn write_into<W: ByteWriter>(&self, target: &mut W) {
+        self.name.write_into(target);
+        self.items.write_into(target);
+    }
+}
+
+impl Deserializable for MetaKeyValue {
+    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
+        let name = Ident::read_from(source)?;
+        let items = BTreeMap::read_from(source)?;
+
+        Ok(Self { span: SourceSpan::UNKNOWN, name, items })
     }
 }