+++ assembly-syntax/src/ast/attribute/meta/list.rs
@@ -1,13 +1,20 @@
 use alloc::vec::Vec;
 
+use miden_core::utils::{
+    ByteReader, ByteWriter, Deserializable, DeserializationError, Serializable,
+};
 use miden_debug_types::{SourceSpan, Spanned};
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 
 use super::MetaExpr;
 use crate::ast::Ident;
 
 /// Represents the metadata of a named list [crate::ast::Attribute], i.e. `@name(item0, .., itemN)`
-#[derive(Clone)]
+#[derive(Debug, Clone)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct MetaList {
+    #[cfg_attr(feature = "serde", serde(skip, default))]
     pub span: SourceSpan,
     /// The identifier used as the name of this attribute
     pub name: Ident,
@@ -100,5 +107,38 @@ impl core::hash::Hash for MetaList {
     fn hash<H: core::hash::Hasher>(&self, state: &mut H) {
         self.name.hash(state);
         self.items.hash(state);
+    }
+}
+
+#[cfg(feature = "arbitrary")]
+impl proptest::arbitrary::Arbitrary for MetaList {
+    type Parameters = ();
+
+    fn arbitrary_with(_args: Self::Parameters) -> Self::Strategy {
+        use proptest::{arbitrary::any, strategy::Strategy};
+
+        let name = any::<Ident>();
+        let items = proptest::collection::vec(any::<MetaExpr>(), 1..3);
+        (name, items)
+            .prop_map(|(name, items)| Self { span: SourceSpan::UNKNOWN, name, items })
+            .boxed()
+    }
+
+    type Strategy = proptest::prelude::BoxedStrategy<Self>;
+}
+
+impl Serializable for MetaList {
+    fn write_into<W: ByteWriter>(&self, target: &mut W) {
+        self.name.write_into(target);
+        self.items.write_into(target);
+    }
+}
+
+impl Deserializable for MetaList {
+    fn read_from<R: ByteReader>(source: &mut R) -> Result<Self, DeserializationError> {
+        let name = Ident::read_from(source)?;
+        let items = Vec::read_from(source)?;
+
+        Ok(Self { span: SourceSpan::UNKNOWN, name, items })
     }
 }