+++ assembly/src/tests.rs
@@ -3291,6 +3291,7 @@ prop_compose! {
                         name: export.name.clone(),
                         digest,
                         signature: export.signature.clone(),
+                        attributes: export.attributes.clone(),
                     });
                 }
             }
@@ -3305,6 +3306,7 @@ prop_compose! {
                     name: main,
                     digest,
                     signature: None,
+                    attributes: Default::default(),
                 });
             }
         }